+++ schema/lvmdb/lvmopsdb.sql
@@ -214,16 +214,6 @@ CREATE TABLE lvmopsdb.guider_coadd (
     exposure_no INTEGER
 );
 
-CREATE TABLE lvmopsdb.overhead (
-    pk SERIAL PRIMARY KEY NOT NULL,
-    observer_id BIGINT,
-    tile_id INTEGER,
-    stage TEXT,
-    start_time DOUBLE PRECISION,
-    end_time DOUBLE PRECISION,
-    duration REAL
-);
-
 
 -- constraints
 
@@ -364,10 +354,6 @@ CLUSTER standard_qc3_index ON lvmopsdb.standard;
 CREATE INDEX CONCURRENTLY ON lvmopsdb.guider_coadd (exposure_no);
 CREATE INDEX CONCURRENTLY ON lvmopsdb.guider_frame (exposure_no);
 CREATE INDEX CONCURRENTLY ON lvmopsdb.agcam_frame (exposure_no);
-
-CREATE INDEX CONCURRENTLY ON lvmopsdb.overhead (observer_id);
-CREATE INDEX CONCURRENTLY ON lvmopsdb.overhead (tile_id);
-CREATE INDEX CONCURRENTLY ON lvmopsdb.overhead (stage);
 
 grant usage on schema lvmopsdb to sdss_user;